+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.2
+
+- Updating colors for quick input
+- Updating colors for welcome page
+
 ## 1.0.0
 
 Adding Commands: